[Feat]뱃지 기능 추가
