+++ src/main/java/com/vlc/maeummal/domain/template/common/dto/TemplateResponseDTO.java
@@ -1,14 +1,11 @@
 package com.vlc.maeummal.domain.template.common.dto;
 
-import com.vlc.maeummal.domain.template.common.TemplateEntity;
+import com.vlc.maeummal.domain.template.common.entity.TemplateEntity;
 import com.vlc.maeummal.global.enums.TemplateType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-import java.util.List;
 
 import static com.vlc.maeummal.global.converter.DateTimeConverter.*;
