+++ src/main/java/com/vlc/maeummal/domain/template/template1/service/Template1Service.java
@@ -1,6 +1,9 @@
 package com.vlc.maeummal.domain.template.template1.service;
 
 import com.vlc.maeummal.domain.lesson.dto.WordDTO;
+import com.vlc.maeummal.domain.member.entity.MemberEntity;
+import com.vlc.maeummal.domain.template.common.entity.BadgeEntity;
+import com.vlc.maeummal.domain.template.common.repository.BadgeRepository;
 import com.vlc.maeummal.domain.template.template1.dto.Template1DTO;
 import com.vlc.maeummal.domain.template.template1.entity.Template1Entity;
 import com.vlc.maeummal.domain.template.template1.repository.Template1Repository;
@@ -13,6 +16,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.lang.reflect.Member;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -23,6 +27,7 @@ public class Template1Service {
     private final Template1Repository template1Repository;
     private final WordRepository wordRepository;
     private final UserAuthorizationConverter userAuthorizationConverter;
+    private final BadgeRepository badgeRepository;
 
     @Transactional
     public Template1DTO createTemplate(String title, Integer level) {
@@ -96,39 +101,6 @@ public class Template1Service {
 
         return convertToDTO(template);
     }
-
-
-//    @Transactional
-//    public Template1DTO addRandomWordsToTemplate(Long templateId) {
-//        Template1Entity template = template1Repository.findById(templateId)
-//                .orElseThrow(() -> new RuntimeException("Template1Entity not found with id: " + templateId));
-//
-//        List<WordEntity> allWords = wordRepository.findAll();
-//        if (allWords.size() < 3) {
-//            throw new RuntimeException("Not enough WordEntities to select 3 random words.");
-//        }
-//
-//        Collections.shuffle(allWords);
-//        List<WordEntity> randomWords = allWords.subList(0, 3);
-//
-//        for (WordEntity word : randomWords) {
-//            // 단어가 이미 다른 템플릿에 연결되어 있는지 확인
-//            if (word.getTemplate1Entity() == null) {
-//                // 양방향 연관 관계 설정
-//                word.setTemplate1Entity(template); // WordEntity에 template1Entity 설정
-//                template.getWordEntities().add(word); // Template1Entity에 WordEntity 추가
-//            } else {
-//                // 이미 다른 템플릿에 연결된 경우 처리 로직 (예: 무시하거나, 로그를 남김)
-//                System.out.println("WordEntity already associated with another template: " + word.getId());
-//            }
-//        }
-//
-//        // template과 word 엔티티들을 데이터베이스에 저장
-//        template1Repository.save(template);
-//        wordRepository.saveAll(randomWords);
-//
-//        return convertToDTO(template);
-//    }
 
     // 내가 만든 템플릿만 보기
     @Transactional
@@ -214,6 +186,5 @@ public class Template1Service {
         // 템플릿 삭제
         template1Repository.delete(template);
     }
-
 
 }
