+++ src/main/java/com/vlc/maeummal/domain/template/template4/entity/Template4Entity.java
@@ -1,7 +1,7 @@
 package com.vlc.maeummal.domain.template.template4.entity;
 
-import com.vlc.maeummal.domain.template.common.TemplateEntity;
-import com.vlc.maeummal.domain.template.common.StoryCardEntity;
+import com.vlc.maeummal.domain.template.common.entity.TemplateEntity;
+import com.vlc.maeummal.domain.template.common.entity.StoryCardEntity;
 import com.vlc.maeummal.global.enums.TemplateType;
 import jakarta.persistence.*;
 import lombok.*;
